Update index.html
While reading through the blog, I found this sentence a bit counter-intuitive and decided to read through the original paper (https://arxiv.org/pdf/2206.04624). According to the last sentence on page 5 and the caption for table 3 on page 6, I believe that the original sentence had the relationship reversed, and that this correction should result in the right relationship.